requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.4
cssselect==1.2.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.6
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-sslserver==0.22
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-social-oauth2==2.1.3
drf-spectacular==0.26.4
drf-yasg==1.21.7
ecdsa==0.18.0
filelock==3.12.4
gevent==23.9.0.post1
greenlet==3.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
kombu==5.3.2
lxml==4.9.3
Markdown==3.4.4
oauth2-provider==0.0
oauthlib==3.2.2
packaging==23.2
parsel==1.8.1
prompt-toolkit==3.0.39
Protego==0.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
Scrapy==2.10.1
scrapy-user-agents==0.1.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tldextract==3.6.0
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.6
user-agents==2.2.0
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.8
whitenoise==6.5.0
wrapt==1.15.0
zope.event==5.0
zope.interface==6.0
